+++ qualinvest_core/src/position.rs
@@ -1,14 +1,14 @@
 use std::sync::Arc;
 
-use finql_data::Currency;
-use finql_postgres::PostgresDB;
+use finql::datatypes::Currency;
+use finql::postgres::PostgresDB;
 use finql::portfolio::{PortfolioPosition, PositionTotals, PositionError, calculate_position_for_period};
 use finql::period_date::PeriodDate;
 
 use crate::accounts::AccountHandler;
 
 // Calculate position for a given period for transactions in a set of accounts
-pub async fn calculate_position_for_period_for_accounts(currency: Currency, account_ids: &[usize], 
+pub async fn calculate_position_for_period_for_accounts(currency: Currency, account_ids: &[i32], 
     start: PeriodDate, end: PeriodDate, db: Arc<PostgresDB>) 
         -> Result<(PortfolioPosition, PositionTotals), PositionError> {
     let end = end.date(None)?;
